:construction: initial refactor
